Extract FileList to a new class and refactor the needed classes
